Flacstream: change from unsigned int to size_t applied to the old API (Debian); changed the name of the flac plugin to reflect the name of the different flac versions.

+++ alsaplayer/input/flac/OggFlacStream.cpp
@@ -188,7 +188,7 @@ OggFlacStream::writeCallBack (const OggFLAC__StreamDecoder * decoder,
 FLAC__StreamDecoderReadStatus
 OggFlacStream::readCallBack (const OggFLAC__StreamDecoder * decoder,
 			     FLAC__byte buffer[],
-			     unsigned * bytes,
+			     size_t * bytes,
 			     void * client_data)
 {
     if (!client_data)
